pantallas crear y editar rest

+++ .NET_Presentacion/PresentacionREST/PresentacionREST/Models/MovimientoDTO.cs
@@ -7,12 +7,12 @@ namespace PresentacionREST.Models
 {
 	public class MovimientoDTO
 	{
-		public decimal id_movimiento { get; set; }
+		public int id_movimiento { get; set; }
 		public Nullable<decimal> num_aprovacion { get; set; }
-		public decimal valor { get; set; }
+		public int valor { get; set; }
 		public System.DateTime fecha { get; set; }
 		public bool estado { get; set; }
-		public decimal id_usuario { get; set; }
+		public int id_usuario { get; set; }
 
 		public virtual UsuarioDTO Usuario { get; set; }
 	}
